Chatbot_Microplstic/requirements.txt
Flask>=2.3.2
google-genai>=0.1.0
requests>=2.31.0
